test(postgres): coude_inventory repo + fix NUMERIC cast bug in claim_primes (+11 tests)
BUG FIX : claim_primes faisait SUM(amount) qui retourne NUMERIC en PostgreSQL, mais le code decodait en Option<i64> -> panic au runtime des qu'il y avait des primes a claim (dette silencieuse). Ajoute ::BIGINT cast sur le COALESCE(SUM). Test prime_claim_returns_total aurait du attraper cela plus tot.

PgCoudeInventoryRepository : items add/use/has/list (quantity accumule via UPSERT), primes create + list_active + claim (SUM total + marque tous comme claimed), insurance buy/get_active/expire (fallback 3600s si duration<=0, false si deja active via INSERT WHERE NOT EXISTS).

+++ services/api/src/adapters/outbound/postgres/coude_inventory_repository.rs
@@ -231,7 +231,7 @@ impl CoudeInventoryRepository for PgCoudeInventoryRepository {
                 WHERE guild_id = $1 AND target_id = $2 AND claimed = FALSE
                 RETURNING amount
             )
-            SELECT COALESCE(SUM(amount), 0) FROM claimed"#,
+            SELECT COALESCE(SUM(amount), 0)::BIGINT FROM claimed"#,
         )
         .bind(guild_id)
         .bind(target_id)
